increase resources available to clinsv to match SLURM request

+++ clinsv/clinsv.run.sh
@@ -5,7 +5,7 @@
 #SBATCH -p icelake,a100cpu
 #SBATCH -N 1
 #SBATCH -n 24
-#SBATCH --time=12:00:00
+#SBATCH --time=48:00:00
 #SBATCH --mem=144GB
 
 # Notification Configuration 
@@ -60,6 +60,12 @@ if [ -z "${Config}" ]; then # If no config file specified use the default
     Config=${scriptDir}/configs/hs38DH.SV_ClinSV.phoenix.cfg
     echo "## INFO: Using the default config ${Config}"
 fi
+if [ ! -f "/hpcfs/groups/phoenix-hpc-neurogenetics/clinsv/phoenix_resources.json" ]; then # Check if the config file exists
+    cp ${defaultResourcesJSON} /hpcfs/groups/phoenix-hpc-neurogenetics/clinsv/phoenix_resources.json
+fi
+if [ ! -d "${neuroDir}/clinsv/test_run" ]; then # Check if the output directory exists
+    mkdir -p ${neuroDir}/clinsv/test_run
+fi
 
 source ${Config}
 
@@ -75,6 +81,9 @@ ${progDir}/${progName} /app/clinsv/bin/clinsv \
 -p /app/project_folder/ \
 -ref /app/ref-data/refdata-b38 \
 -i "/app/input/*.bam" \
+-j /app/project_folder/phoenix_resources.json \
 -w 
 
-rm $neuroDir/clinsv/clinsv.lock # Remove the lock file to allow other runs to proceed
+if [ -f "${neuroDir}/clinsv/clinsv.lock" ]; then # Check if the lock file exists
+    rm $neuroDir/clinsv/clinsv.lock # Remove the lock file to allow other runs to proceed
+fi
